backend/pyproject.toml
[project]
name = "a-erp-backend"
version = "2.0.0"
description = "A-ERP API — Agentic Enterprise Resource Platform (FastAPI + PostgreSQL)"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.111.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "slowapi>=0.1.9",
    "redis>=5.0.0",
    "arq>=0.25.0",
    "structlog>=24.0.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.1",
    "email-validator>=2.1.0",
    "greenlet>=3.0.0",
    "bcrypt>=5.0.0",
    "croniter>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "factory-boy>=3.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]
